chore: checks that run, rather than claims that were written down
The design system said contrast was verified by a script in CI, that axe-core
ran in CI, and that Lighthouse hit 95 and 100. None of that was true. There was
no CI. So the checks now exist and run.

tools/a11y_audit.py loads every page in both themes at 390, 768 and 1280 pixels
wide, which is 30 page renders, and fails on any WCAG 2.2 A or AA violation,
any page that scrolls sideways, any control under its target size, or any
console error. It found real problems: code blocks that scroll but cannot take
keyboard focus, two links in a table cell closer together than 24 pixels, info
buttons at 20 by 20 against a WCAG 2.2 minimum of 24, and radio labels and
number fields under the size the design system claims for a product screen. All
fixed. Both projects are now clean.

tools/check_copy.py enforces the two house rules, no emoji and no em or en
dashes, which were being kept by hand and had drifted into both video scripts.

.github/workflows/ci.yml runs the tests, ruff, the copy check and the
accessibility audit on every push. Six tests that drive a real model or stand
up real A2A servers are marked `aws` and deselected there rather than made to
pass against a fake; the other 122 run.

ruff had never actually been configured, so it was running on defaults and
reporting 83 findings. It now has an explicit rule set with the ignores written
down and argued: naive datetimes are deliberate because the clock is simulated,
and a broad except around AWS is the fallback path working as intended. Every
remaining finding is fixed, including five zip() calls that would have paired
mismatched lists in silence.

Also: the deploy scripts that actually produced the live service are now in the
repository, README documents deployment including the App Runner digest trap
that cost an afternoon, and the test counts and the AgentCore claims in
DEMO_AND_VIDEO.md are corrected to what the system does.

+++ src/turnout/agents/peer_service.py
@@ -33,7 +33,7 @@ def travel_minutes(a: tuple[float, float], b: tuple[float, float]) -> int:
     dlat, dlon = lat2 - lat1, lon2 - lon1
     h = math.sin(dlat / 2) ** 2 + math.cos(lat1) * math.cos(lat2) * math.sin(dlon / 2) ** 2
     km = 2 * 6371 * math.asin(math.sqrt(h))
-    return int(round(km / SPEED_KMH * 60)) + TURNOUT_MIN
+    return round(km / SPEED_KMH * 60) + TURNOUT_MIN
 
 
 def evaluate_request(req: CoverageRequest, rt=None) -> CoverageOffer:
